Deployment/app/requirements.txt
altair==5.4.0
amqp==5.2.0
async-timeout==4.0.3
attrs==24.2.0
billiard==4.2.0
blinker==1.8.2
cachetools==5.5.0
celery==5.4.0
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
exceptiongroup==1.2.2
flower==2.0.1
gitdb==4.0.11
GitPython==3.1.43
h11==0.14.0
humanize==4.10.0
idna==3.8
Jinja2==3.1.4
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
kombu==5.4.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
narwhals==1.5.5
numpy==2.1.0
outcome==1.3.0.post0
packaging==24.1
pandas==2.2.2
pillow==10.4.0
prometheus_client==0.20.0
prompt_toolkit==3.0.47
protobuf==5.27.3
pyarrow==17.0.0
pydeck==0.9.1
Pygments==2.18.0
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-decouple==3.8
pytz==2024.1
redis==5.0.8
referencing==0.35.1
requests==2.32.3
rich==13.8.0
rpds-py==0.20.0
selenium==4.23.1
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
sortedcontainers==2.4.0
streamlit==1.37.1
tenacity==8.5.0
toml==0.10.2
tornado==6.4.1
trio==0.26.2
trio-websocket==0.11.1
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
vine==5.1.0
watchdog==4.0.2
wcwidth==0.2.13
websocket-client==1.8.0
wsproto==1.2.0
